Add early support for 1D and 3D model fitting.
