+++ baccmod/__init__.py
@@ -6,6 +6,7 @@
 # This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
 # ---------------------------------------------------------------------
 
+
 from importlib import import_module
 
 __all__ = [
